requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.1
blinker==1.4
blis==0.4.1
boto3==1.18.5
botocore==1.21.5
breadability==0.1.20
cachetools==4.2.2
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
debugpy==1.4.0
decorator==5.0.9
defusedxml==0.7.1
docopt==0.6.2
en-core-web-sm==2.3.1
entrypoints==0.3
enum-compat==0.0.3
gensim==3.8.3
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.3.2
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.5
notebook==6.4.0
numpy==1.18.5
packaging==21.0
pandas==1.1.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
plac==1.1.3
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyarrow==4.0.1
pycountry==20.7.3
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.1.0
rake-nltk==1.0.4
regex==2021.7.6
requests==2.26.0
s3transfer==0.5.0
scikit-learn==0.24.1
scipy==1.7.0
seaborn==0.11.0
Send2Trash==1.7.1
six==1.16.0
smart-open==5.1.0
smmap==4.0.0
spacy==2.3.2
spacy-streamlit==0.1.0
srsly==1.0.5
streamlit==0.69.2
sumy==0.8.1
terminado==0.10.1
testpath==0.5.0
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.0
